@@ -1,16 +1,17 @@
 name: html_text
 description: convert html text to text by putting italics, bold and underline
-version: 0.0.3
-homepage: https://github.com/matheusperez/html_text
+version: 0.0.5
+homepage: https://github.com/gustavorw/html_text
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: ">=2.14.0 <3.0.0"
+  flutter: ">=2.5.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  google_fonts: ^1.1.2
+  google_fonts: 2.1.0
+  cached_network_image: ^3.2.0
 
 dev_dependencies:
   flutter_test:
